keibaai/gui/requirements.txt
# KeibaAI_v2 GUI Requirements
# Streamlit-based web interface dependencies

# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional but recommended
watchdog>=3.0.0  # For auto-reload
